Fixed a signed arithmetic bug in minor-gc.c:91.
git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/trunk@1642 ac4fbcad-d21d-0410-b760-c7b1d3380187

+++ src/lib/parallel-rt/gc/minor-gc.c
@@ -88,7 +88,7 @@ void MinorGC (VProc_t *vp, Value_t **roots)
 	if (isMixedHdr(hdr)) {
 	  // a record
 	    Word_t tagBits = GetMixedBits(hdr);
-	    assert (tagBits < (1 << GetMixedSizeW(hdr)));
+	    assert ((uint64_t)tagBits < (1l << (uint64_t)GetMixedSizeW(hdr)));
 	    Value_t *scanP = (Value_t *)nextScan;
 	    while (tagBits != 0) {
 		if (tagBits & 0x1) {
